backend/requirements.txt
flask
flask-cors
google-genai
python-dotenv
pillow
requests
gunicorn
pymongo
